requirements.txt
aiohttp==3.7.4.post0
asgiref==3.5.2
async-timeout==3.0.1
attrs==21.4.0
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.1.0
discord.py==1.7.3
Django==4.0.5
django-dotenv==1.4.2
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
multidict==6.0.2
oauthlib==3.2.0
psycopg2==2.9.3
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
typing_extensions==4.2.0
urllib3==1.26.10
yarl==1.7.2
